backend/requirements.txt
flask
flask-cors
torch
torchvision
torchaudio
numpy
pillow
pyyaml
